Add files via upload
This code is to help create visual maps in powerBI using the BING map visual which only takes lat & long column data in powerBI. This R script interprets columns which are named as easting and northing in datasets such as the CAMS ledger and converts them into new columns for lat & long. It can be run as an R script within powerBI provided that R is installed locally as well as PowerBI desktop. It was generated quickly with the help of ChatGPT